Scheduling rewrites of the h1 tag changed to correct an animation glitch when returning to the tab after beeing in the background.
Chrome slows down setIntervals that run more frequently than once per
second when a tab is in the background or the browser is minimized. This
causes rewrite animation overlapping when returning to the tab after
beeing in the background for some time. To avoid it,, now a rewrite is
scheduled with setInterval when the previous rewrite completes instead
on relying on setInterval.